+++ CHANGELOG.md
@@ -21,12 +21,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Breaking Changes
 
-## zap [0.2.1] [PR #15](https://github.com/MeteoraAg/zap-program/pull/41)
+## zap [0.2.1] [PR #41](https://github.com/MeteoraAg/zap-program/pull/41)
+
+### Added
+
+- Added dammv1 swap as a zap out option
 
 ### Fixed
 
 - Fix zap in damm-v2 with new base fee mode
-- Add dammv1 swap as a zap out option
 
 ## zap [0.2.0] [PR #15](https://github.com/MeteoraAg/zap-program/pull/15)